requirements.txt
flask
flask-cors
bootstrap-flask
gunicorn
requests
numpy
pillow
torch 
torchvision

